Fix font loading for overlay
